refactor: 게시글 필터 상수 정리 #52
